[random] move randomgenerator to its own pkg to avois cycle dependencies between namer and utils
